Add files via upload
The continuous features, **Score** and **Comments Count**, were not found suitable for predicting the Flair because of the percentage of outliers in their data points for each Flair. **Time Stamp** feature was also dropped, after an attempt to find a trend between the number of posts of each Flair posted in a particular month, no such trend was found.